backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
celery>=5.3.4
redis>=5.0.1
django-redis>=5.4.0
djangorestframework-simplejwt>=5.3.1
Pillow>=10.1.0
python-dateutil>=2.8.2
pytz>=2023.3
drf-spectacular>=0.27.0
django-allauth>=0.57.0
django-guardian>=2.4.0
django-role-permissions>=3.0.0


